Renamed and changed year inputs to number
